Created contact us page
Created contact us page and started the css for the page, I want to add php so that it functions correctly so will start to learn that.

In terms of the css file, try to get all of it on one file and section it off using comments.

also created a folder for images so please put images in that folder and when linking use "images/filename" to add it to the site.